requirements.txt
Flask
bs4
lxml
requests
Flask-SQLAlchemy
Pillow